+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - [v0.10.0](#0100)
   - [v0.10.1](#0101)
   - [v0.10.2](#0102)
+  - [v0.10.3](#0103)
 
 
 ## 0.1.0
@@ -211,3 +212,7 @@ Rewritten every thing as a module with `nmanga` namespace.
   - To enable, use `nmanga config` and configure the experimental part
 - Raw tagging with provided metadata instead of custom metadata formatting we use
 - Add support for special numbered volume (v01.5) etc.
+
+### 0.10.3
+**Fixes**
+- Properly support special volume scanning/collection